FolderNameFilter: allow multiple tags per folder
So far, the last entry wins when a folder is specified multiple times as
a folder transform. The reason is that the folder name is a key in a
dict whose values are tags. On the other hand, if a message has copies
in several folders then all these rules apply.

Change (fix?) afew's FolderNameFilter so that multiple transforms for
the same folder lead to multiple tags being added. Technically, this is
done by using sets as values in the transform dict.

In the future, we might think about implementing notmuch-tag style
notation (with the obvious quoting issues).